+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Show UTC datetime when hovering a datetime
 - Implement Dark mode and theme switcher
 - Create a settings menu (#157)
+- (docs): added recipe SSL certificate check
 
 ### Fixed
 - Sometimes logs are appended to an existing logs files of previous runs (#131)